Adding log content tests
