+++ tests/CloudFoundry/steps/CloudController.py
@@ -208,6 +208,11 @@ def step_impl(context):
     context.number_of_results = logs.get('total_results')
 
 
+@when('I look at the audit logs')
+def step_impl(context):
+    context.sample_log = context.user.events()['resources'][0]
+
+
 # Thens
 @then('the org exists')
 def step_impl(context):
@@ -324,3 +329,25 @@ def step_impl(context, number):
     print(context.number_of_results)
     assert int(number) == context.number_of_results
 
+
+@then('audit logs have timestamp')
+def step_impl(context):
+    assert 'created_at' in context.sample_log['metadata']
+
+
+@then('audit logs have type of event')
+def step_impl(context):
+    print(context.sample_log)
+    assert 'type' in context.sample_log['entity']
+
+
+@then('audit logs have actor')
+def step_impl(context):
+    print(context.sample_log)
+    assert 'actor' in context.sample_log['entity']
+
+
+@then('audit logs have actee')
+def step_impl(context):
+    print(context.sample_log)
+    assert 'actee' in context.sample_log['entity']
