+++ tests/CloudFoundry/steps/UAA.py
@@ -1,7 +1,6 @@
-import datetime
 import config
 
-from behave import given, when, then
+from behave import when, then
 
 from cloudfoundry import Client
 
@@ -23,34 +22,6 @@ def step_impl(context, times):
             verify_ssl=False
         )
         assert not user.is_logged_in()
-        
-
-@when('I look at the audit logs')
-def step_impl(context):
-    context.sample_log = context.user.events()['resources'][0]
-
-
-@then('audit logs have timestamp')
-def step_impl(context):
-    assert 'created_at' in context.sample_log['metadata']
-
-
-@then('audit logs have type of event')
-def step_impl(context):
-    print(context.sample_log)
-    assert 'type' in context.sample_log['entity']
-
-
-@then('audit logs have actor')
-def step_impl(context):
-    print(context.sample_log)
-    assert 'actor' in context.sample_log['entity']
-
-
-@then('audit logs have actee')
-def step_impl(context):
-    print(context.sample_log)
-    assert 'actee' in context.sample_log['entity']
 
 
 @then('I am locked out')
